+++ saw-remote-api/Main.hs
@@ -46,6 +46,8 @@ import SAWServer.MIRCrucibleSetup
     ( mirLoadModuleDescr, mirLoadModule )
 import SAWServer.MIRFindADT
     ( mirFindADTDescr, mirFindADT )
+import SAWServer.MIRFindMangledADT
+    ( mirFindMangledADTDescr, mirFindMangledADT )
 import SAWServer.MIRVerify
     ( mirAssumeDescr, mirAssume,
       mirVerifyDescr, mirVerify )
@@ -163,6 +165,10 @@ sawMethods =
      "SAW/MIR/find ADT"
      mirFindADTDescr
      mirFindADT
+  , Argo.command
+     "SAW/MIR/find mangled ADT"
+     mirFindMangledADTDescr
+     mirFindMangledADT
   -- Yosys
   , Argo.command
      "SAW/Yosys/import"