+++ CHANGES.md
@@ -70,6 +70,14 @@ This release supports [version
   Accordingly, the second and third `String` arguments to
   `write_coq_cryptol_primitives_for_sawcore` have been removed.
 
+## New Features
+
+* Add a `mir_find_mangled_adt : MIRModule -> String -> MIRAdt` command, which
+  allows looking up MIR ADTs by their full, mangled names.
+  `mir_find_mangled_adt`'s use is discouraged in favor of using `mir_find_adt`
+  instead, but `mir_find_mangled_adt` can be useful in scenarios where
+  `mir_find_adt` isn't expressive enough to look up particular ADTs (e.g., ADTs
+  that use const generics).
 
 # Version 1.4 -- date still TBD